crond: add job execution that are not related to a username
Until now it was not possible to use crond jobs independent of a
deticated username on the system. In the system there are normaly also
other crond jobs that need to be processed.

For this task you have to install an additional crond daemon to run these
tasks or add all that jobs that need to be done to the root crontab file.

On an embeded system where there is usually only root, all services have
to edit or change this file. This is difficult and unfamiliar users may
accidentally delete these tasks. Since they have not created them as
root.

So that this is no longer necessary, each service can be great their tasks
file and the crond executes these tasks as root in the file. If the
name of the file does not have an deticated user on the system.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>